% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{load.pack1}
\alias{load.pack1}
\title{load.pack1}
\usage{
load.pack1(...)
}
\arguments{
\item{packages}{pre-define packages list.}
}
\description{
To install require packages. Use ipak function to install desired packages.
}
\examples{
load.pack1()
}
\keyword{load.pack1}
