% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{step1}
\alias{step1}
\title{step1}
\usage{
step1(working.folder)
}
\arguments{
\item{working.folder}{mandatory. ex:working.folder<-"C:/Users/lpheng/Desktop/xptdefine".}
}
\description{
step1 funtion will map the working folder and create excel spreadsheet file that needed to be
be completed before running step2.
}
\examples{
step1(working.folder)
}
\keyword{step1}
