% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{step2}
\alias{step2}
\title{step2}
\usage{
step2()
}
\description{
Create subfolders, import files in the list of files.csv to input folder and create studydefinelist.csv.
Note that a WARNING message may appear when one or more labels are longer than 8 characters. An excel file called Var_name_GT8.csv
will be generated for these labels. You can change the label to appropriate length. Note also that you can flag variables to be remove in the excel spreadsheets.
}
\examples{
step2()

}
\keyword{step2}
