% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{step3}
\alias{step3}
\title{Step 3
Optional. Run this option to apply the changes to label name or lable removal.}
\usage{
step3()
}
\description{
Step 3
Optional. Run this option to apply the changes to label name or lable removal.
}
\examples{
step3()
}
\keyword{step3}
