% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{step4a}
\alias{step4a}
\title{Step 4a}
\usage{
step4a(title = "Title")
}
\arguments{
\item{title}{Project title}

\item{logo}{Add logo by copying logo.jpg to the main working folderfirst time package user, run this function to install all required packages.
Can be downloaded from my Github}

\item{style}{Load document style by copying word template and rename as style.docx to the main working folder.
Can be downloaded from my Github}
}
\description{
Convert csv to xpt and create define document
}
\examples{
step4a(title="Add title here")
}
\keyword{step4}
