% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{vardefine}
\alias{vardefine}
\title{vardefine}
\usage{
vardefine(data, maxlevel = 7, vartype = T, labels = FALSE,
  digits = max(3, getOption("digits") - 3), exp.csv = FALSE)
}
\description{
Internal use.
}
\examples{
vardefine()
}
\keyword{vardefine}
